api/requirements.txt
anthropic==0.39.0
pdfplumber==0.11.0
python-docx==1.1.0
Pillow==10.2.0
